requirements.txt
pandas 
matplotlib.pyplot 
seaborn 
numpy
